requirements.txt
anyio==4.6.2.post1
certifi==2025.10.5
charset-normalizer==3.4.4
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.11
pillow==11.0.0
pip==25.3
psycopg2-binary==2.9.11
pyTelegramBotAPI==4.29.1
python-dotenv==1.0.1
python-telegram-bot==21.7
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
telebot==0.0.5
urllib3==2.5.0
